+++ src/TortoiseProc/RebaseDlg.cpp
@@ -1,6 +1,6 @@
 ﻿// TortoiseGit - a Windows shell extension for easy version control
 
-// Copyright (C) 2008-2022 - TortoiseGit
+// Copyright (C) 2008-2023 - TortoiseGit
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -1208,11 +1208,21 @@ void CRebaseDlg::RewriteNotes()
 	if (rewrites.IsEmpty())
 		return;
 	CString tmpfile = GetTempFile();
+	if (tmpfile.IsEmpty())
+	{
+		MessageBox(L"Could not create temp file.", L"TortoiseGit", MB_OK | MB_ICONERROR);
+		return;
+	}
 	tmpfile.Replace(L'\\', L'/');
 	if (!CStringUtils::WriteStringToTextFile(tmpfile, rewrites))
 		return;
 	SCOPE_EXIT{ ::DeleteFile(tmpfile); };
 	CString pipefile = GetTempFile();
+	if (pipefile.IsEmpty())
+	{
+		MessageBox(L"Could not create temp file.", L"TortoiseGit", MB_OK | MB_ICONERROR);
+		return;
+	}
 	pipefile.Replace(L'\\', L'/');
 	CString pipecmd;
 	pipecmd.Format(L"git notes copy --for-rewrite=rebase < %s", static_cast<LPCWSTR>(tmpfile));
@@ -1470,6 +1480,11 @@ void CRebaseDlg::OnBnClickedContinue()
 				return;
 			}
 			CString tempfile = ::GetTempFile();
+			if (tempfile.IsEmpty())
+			{
+				MessageBox(L"Could not create temp file.", L"TortoiseGit", MB_OK | MB_ICONERROR);
+				return;
+			}
 			SCOPE_EXIT{ ::DeleteFile(tempfile); };
 			if (CAppUtils::SaveCommitUnicodeFile(tempfile, str))
 			{
@@ -1606,6 +1621,11 @@ void CRebaseDlg::OnBnClickedContinue()
 		}
 
 		CString tempfile=::GetTempFile();
+		if (tempfile.IsEmpty())
+		{
+				MessageBox(L"Could not create temp file.", L"TortoiseGit", MB_OK | MB_ICONERROR);
+				return;
+		}
 		SCOPE_EXIT{ ::DeleteFile(tempfile); };
 		if (CAppUtils::SaveCommitUnicodeFile(tempfile, str))
 		{
